requirements.txt
# Telegram Bot
aiogram==3.4.1
aiohttp==3.9.1

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.1
